+++ src/common/command/pods.h
@@ -28,7 +28,7 @@ struct tInfo {
 /** \brief Struct data for get_me and get_accout command */
 struct UserInfoData {
     UserInfoData() = default;
-    UserInfoData(uint16_t abank_, uint32_t auser_, uint16_t bbank_, uint16_t buser_, uint32_t time_)
+    UserInfoData(uint16_t abank_, uint32_t auser_, uint16_t bbank_, uint32_t buser_, uint32_t time_)
         : abank{abank_},
           auser{auser_},
           bbank{bbank_},
@@ -51,7 +51,7 @@ struct UserInfoData {
 /** \brief Full struct data for get_me and get_accout command with signature */
 struct UserInfo {
     UserInfo() = default;
-    UserInfo(uint16_t abank, uint32_t auser, uint16_t bbank, uint16_t buser, uint32_t time)
+    UserInfo(uint16_t abank, uint32_t auser, uint16_t bbank, uint32_t buser, uint32_t time)
         : info(abank, auser, bbank, buser, time) {
     }
